+++ src/main/webapp/components/index-cards/index-cards.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { storiesOf } from '../@storybook/react'
+import { storiesOf } from '../../@storybook/react'
 
 import { number } from '@connexta/ace/@storybook/addon-knobs'
