+++ src/main/webapp/components/network-retry/index.tsx
@@ -1,0 +1,4 @@
+import InlineRetry from './inline-retry'
+const SnackbarRetry = require('./snackbar-retry')
+
+export { InlineRetry, SnackbarRetry }
