+++ src/main/webapp/components/network-retry/snackbar-retry.stories.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import { storiesOf } from '../@storybook/react'
-const stories = storiesOf('Retry', module)
-import RetryNotification from './retry'
+import { storiesOf } from '../../@storybook/react'
+import RetryNotification from './snackbar-retry'
 import { text } from '@connexta/ace/@storybook/addon-knobs'
 import { action } from '@connexta/ace/@storybook/addon-actions'
 
-stories.add('Retry', () => {
+const stories = storiesOf('Network Retry', module)
+stories.add('Snackbar', () => {
   const message = text(
     'Message',
     'Issue retrieving [thing], would you like to retry?'
