+++ src/main/webapp/search/details.js
@@ -9,7 +9,7 @@ import { getIn } from 'immutable'
 import React from 'react'
 import { useParams } from 'react-router-dom'
 import Thumbnail from '../thumbnail/thumbnail'
-import ErrorMessage from '../error'
+import ErrorMessage from '../components/network-retry/inline-retry'
 
 const LoadingComponent = () => <LinearProgress />
