+++ src/main/webapp/search/results.js
@@ -13,7 +13,7 @@ import ImageIcon from '@material-ui/icons/Image'
 import TextFieldsIcon from '@material-ui/icons/TextFields'
 import gql from 'graphql-tag'
 import React from 'react'
-import ErrorMessage from '../error'
+import ErrorMessage from '../components/network-retry/inline-retry'
 import Thumbnail from '../thumbnail/thumbnail'
 import Search from './search'
 import useSearchRouting, { DetailsLink } from './search-routing'
