+++ src/main/webapp/user-settings/result-form-select.js
@@ -7,7 +7,7 @@ import React, { useEffect } from 'react'
 import OutlinedSelect from '../input/outlined-select'
 import { useApolloFallback } from '../react-hooks'
 import LinearProgress from '@material-ui/core/LinearProgress'
-import ErrorMessage from '../error'
+import ErrorMessage from '../components/network-retry/inline-retry'
 
 const resultForms = gql`
   query SearchFormsAttributeGroup {
